+++ stroom-data/stroom-data-store-util/src/main/java/stroom/data/store/util/ToolModule.java
@@ -26,12 +26,9 @@ import stroom.statistics.mock.MockInternalStatisticsModule;
 import stroom.task.mock.MockTaskModule;
 import stroom.util.db.ForceLegacyMigration;
 import stroom.util.entityevent.EntityEventBus;
-import stroom.util.io.HomeDirProvider;
-import stroom.util.io.HomeDirProviderImpl;
+import stroom.util.io.DirProvidersModule;
 import stroom.util.io.PathConfig;
 import stroom.util.io.StroomPathConfig;
-import stroom.util.io.TempDirProvider;
-import stroom.util.io.TempDirProviderImpl;
 import stroom.util.servlet.MockServletModule;
 
 import com.google.inject.AbstractModule;
@@ -63,8 +60,7 @@ public class ToolModule extends AbstractModule {
         });
 
         bind(PathConfig.class).to(StroomPathConfig.class);
-        bind(HomeDirProvider.class).to(HomeDirProviderImpl.class);
-        bind(TempDirProvider.class).to(TempDirProviderImpl.class);
+        install(new DirProvidersModule());
     }
 
     @Provides