+++ stroom-headless/src/main/java/stroom/headless/CliModule.java
@@ -34,13 +34,10 @@ import stroom.statistics.api.InternalStatisticsReceiver;
 import stroom.task.impl.TaskContextModule;
 import stroom.util.entityevent.EntityEventBus;
 import stroom.util.io.BasicStreamCloser;
-import stroom.util.io.HomeDirProvider;
-import stroom.util.io.HomeDirProviderImpl;
+import stroom.util.io.DirProvidersModule;
 import stroom.util.io.PathConfig;
 import stroom.util.io.StreamCloser;
 import stroom.util.io.StroomPathConfig;
-import stroom.util.io.TempDirProvider;
-import stroom.util.io.TempDirProviderImpl;
 import stroom.util.pipeline.scope.PipelineScopeModule;
 import stroom.util.pipeline.scope.PipelineScoped;
 import stroom.util.servlet.MockServletModule;
@@ -104,8 +101,7 @@ public class CliModule extends AbstractModule {
         bind(StreamCloser.class).to(BasicStreamCloser.class).in(PipelineScoped.class);
 
         bind(PathConfig.class).to(StroomPathConfig.class);
-        bind(HomeDirProvider.class).to(HomeDirProviderImpl.class);
-        bind(TempDirProvider.class).to(TempDirProviderImpl.class);
+        install(new DirProvidersModule());
     }
 
     @Provides