+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/ProxyPathConfig.java
@@ -3,6 +3,7 @@ package stroom.proxy.app;
 import stroom.util.config.annotations.ReadOnly;
 import stroom.util.io.PathConfig;
 import stroom.util.shared.IsProxyConfig;
+import stroom.util.shared.validation.ValidDirectoryPath;
 
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
@@ -16,6 +17,7 @@ public class ProxyPathConfig extends PathConfig implements IsProxyConfig {
 
     @Override
     @ReadOnly
+    @ValidDirectoryPath
     @JsonPropertyDescription("By default, unless configured otherwise, all other configured paths " +
             "(except proxyConfig.path.temp) will be relative to this directory. If this value is null then" +
             "Stroom-Proxy will use either of the following to derive proxyConfig.path.home: the directory of the " +
@@ -28,6 +30,7 @@ public class ProxyPathConfig extends PathConfig implements IsProxyConfig {
     @Override
     @Nonnull
     @ReadOnly
+    @ValidDirectoryPath
     @JsonPropertyDescription("This directory is used by stroom-proxy to write any temporary file to. " +
             "Should only be set per node in application YAML configuration file.")
     public String getTemp() {