+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/handler/ForwardDestinationConfig.java
@@ -2,12 +2,15 @@ package stroom.proxy.app.handler;
 
 import stroom.util.cert.SSLConfig;
 import stroom.util.shared.AbstractConfig;
+import stroom.util.shared.IsProxyConfig;
 import stroom.util.shared.NotInjectableConfig;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.util.Objects;
+
 @NotInjectableConfig
-public class ForwardDestinationConfig extends AbstractConfig {
+public class ForwardDestinationConfig extends AbstractConfig implements IsProxyConfig {
 
     private String forwardUrl;
     private Integer forwardTimeoutMs = 30000;
@@ -75,5 +78,36 @@ public class ForwardDestinationConfig extends AbstractConfig {
     @JsonProperty
     public void setSslConfig(final SSLConfig sslConfig) {
         this.sslConfig = sslConfig;
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        final ForwardDestinationConfig that = (ForwardDestinationConfig) o;
+        return Objects.equals(forwardUrl, that.forwardUrl) && Objects.equals(forwardTimeoutMs,
+                that.forwardTimeoutMs) && Objects.equals(forwardDelayMs,
+                that.forwardDelayMs) && Objects.equals(forwardChunkSize,
+                that.forwardChunkSize) && Objects.equals(sslConfig, that.sslConfig);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(forwardUrl, forwardTimeoutMs, forwardDelayMs, forwardChunkSize, sslConfig);
+    }
+
+    @Override
+    public String toString() {
+        return "ForwardDestinationConfig{" +
+                "forwardUrl='" + forwardUrl + '\'' +
+                ", forwardTimeoutMs=" + forwardTimeoutMs +
+                ", forwardDelayMs=" + forwardDelayMs +
+                ", forwardChunkSize=" + forwardChunkSize +
+                ", sslConfig=" + sslConfig +
+                '}';
     }
 }