+++ stroom-proxy/stroom-proxy-app/src/main/java/stroom/proxy/app/handler/ForwardStreamConfig.java
@@ -1,5 +1,6 @@
 package stroom.proxy.app.handler;
 
+import stroom.util.shared.AbstractConfig;
 import stroom.util.shared.IsProxyConfig;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -11,7 +12,7 @@ import javax.inject.Singleton;
 
 @Singleton
 @JsonPropertyOrder(alphabetic = true)
-public class ForwardStreamConfig implements IsProxyConfig {
+public class ForwardStreamConfig extends AbstractConfig implements IsProxyConfig {
 
     private boolean isForwardingEnabled = false;
     private String userAgent;