+++ stroom-util-shared/src/main/java/stroom/util/shared/AbstractConfig.java
@@ -5,11 +5,10 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.util.Objects;
 
 /**
- * Marker interface for java bean classes used to provide configuration properties.
- * This includes AppConfig and the classes that sit beneath it.  Implementing classes
- * are expected to be (de)serialised from/to YAML configuration files.
+ * Super class for all stroom config pojos. Can be decorated with property path
+ * information, e.g. stroom.path.home
  */
-public abstract class AbstractConfig {
+public abstract class AbstractConfig implements HasPropertyPath {
 
     // Held in part form to reduce memory overhead as some parts will be used
     // many times over all the config objects
@@ -19,6 +18,7 @@ public abstract class AbstractConfig {
     /**
      * @return The base property path, e.g. "stroom.node" for this config object
      */
+    @Override
     @JsonIgnore
     public String getBasePath() {
         Objects.requireNonNull(basePropertyPath);
@@ -29,14 +29,17 @@ public abstract class AbstractConfig {
      * @return The full property path, e.g. "stroom.node.status" for the named property on this config
      * object
      */
+    @Override
     public String getFullPath(final String propertyName) {
         Objects.requireNonNull(basePropertyPath);
         Objects.requireNonNull(propertyName);
         return basePropertyPath.merge(propertyName).toString();
     }
 
+    @Override
     @JsonIgnore
     public void setBasePath(final PropertyPath basePropertyPath) {
         this.basePropertyPath = basePropertyPath;
     }
+
 }