+++ stroom-util/src/main/java/stroom/util/validation/ValidRegexValidatorImpl.java
@@ -1,16 +1,27 @@
-package stroom.config.global.impl.validation;
+package stroom.util.validation;
 
-import stroom.util.scheduler.MalformedCronException;
-import stroom.util.scheduler.SimpleCron;
-import stroom.util.shared.validation.ValidSimpleCron;
-import stroom.util.shared.validation.ValidSimpleCronValidator;
+import stroom.util.shared.validation.ValidRegex;
+import stroom.util.shared.validation.ValidRegexValidator;
 
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
 import javax.validation.ConstraintValidatorContext;
 
-public class ValidSimpleSimpleCronValidatorImpl implements ValidSimpleCronValidator {
+public class ValidRegexValidatorImpl implements ValidRegexValidator {
 
+    /**
+     * Initializes the validator in preparation for
+     * {@link #isValid(Object, ConstraintValidatorContext)} calls.
+     * The constraint annotation for a given constraint declaration
+     * is passed.
+     * <p/>
+     * This method is guaranteed to be called before any use of this instance for
+     * validation.
+     *
+     * @param constraintAnnotation annotation instance for a given constraint declaration
+     */
     @Override
-    public void initialize(final ValidSimpleCron constraintAnnotation) {
+    public void initialize(final ValidRegex constraintAnnotation) {
 
     }
 
@@ -31,8 +42,8 @@ public class ValidSimpleSimpleCronValidatorImpl implements ValidSimpleCronValida
 
         if (value != null) {
             try {
-                SimpleCron.compile(value);
-            } catch (MalformedCronException e) {
+                Pattern.compile(value);
+            } catch (PatternSyntaxException e) {
                 final String msgTemplate =
                         context.getDefaultConstraintMessageTemplate() +
                                 ". caused by: " +