+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#1469** : Add hot loading of config to proxy.
+
+* Change proxy and stroom config validation to cope with relative paths and `~`.
+  
+* Issue **#2353** : Swallow NoSuchFileException in config monitor.
+
 
 ## [v7.0-beta.139] - 2021-08-17